memara/ Minor : cleaning and arranging files - adjust readme file for reviews
