+++ README.md
@@ -1,0 +1,54 @@
+🎙️ Egyptian Arabic Speech Recognition App
+
+Bringing Accurate and Real-Time Egyptian Arabic Speech-to-Text to Mobile
+
+🚀 This mobile app integrates a custom Egyptian Arabic Speech Recognition Model, enabling seamless and precise speech-to-text conversion in real-time. Whether you're taking notes, sending messages, or transcribing audio, this app delivers a smooth and efficient experience.
+
+🔥 Key Features
+
+✅ Real-Time Speech Recognition – Convert spoken Egyptian Arabic into text instantly.✅ High Accuracy – Powered by a custom-trained model optimized for Egyptian dialect. ✅ Intuitive UI – Minimalistic and user-friendly design for effortless usage.✅ Export & Share – Copy, save, or share transcriptions with just a tap.✅ Lightweight & Fast – Optimized for performance and efficiency.
+
+📸 Screenshots
+
+📥 Installation & Setup
+
+🔧 Prerequisites
+
+Ensure you have Flutter installed on your system. If not, follow the official Flutter installation guide.
+
+🏗️ Build & Run the App
+
+git clone https://github.com/YOUR-USERNAME/YOUR-REPO.git
+cd YOUR-REPO
+flutter pub get
+flutter run
+
+💡 Note: If any API keys or configuration files are required, follow the setup instructions in the config/ folder.
+
+🛠️ Usage Guide
+
+1️⃣ Open the app and tap the microphone button.2️⃣ Speak in Egyptian Arabic, and watch the app transcribe your words in real-time.3️⃣ Tap copy or share to export your text.4️⃣ Customize settings (e.g., dark mode, sensitivity) in the Settings section.
+
+🧠 About the Speech Recognition Model
+
+🔹 Custom-built for Egyptian Arabic – Trained on a dataset specifically tailored to local dialects.🔹 Optimized for speed & accuracy – Delivers quick and precise results.🔹 Designed for mobile devices – state-of-art performance on the EACSC dataset.
+
+🤝 Contributing
+
+💡 Want to improve this project? Contributions are welcome!
+
+Report bugs by opening an issue.
+
+Suggest new features via pull requests.
+
+Enhance the model by contributing better training data.
+
+📜 License
+
+📝 This project is licensed under the MIT License – see the LICENSE file for details.
+
+📩 Contact & Support
+
+💬 Have questions or need support? Feel free to reach out:📧 Email: your-email@example.com🔗 LinkedIn: Your LinkedIn🐦 Twitter: @YourTwitter
+
+🌟 If you find this project useful, don’t forget to ⭐ star the repository! 🚀
