+++ speech_to_text_app/lib/main.dart
@@ -1,8 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_sound/flutter_sound.dart';
-import 'package:path_provider/path_provider.dart';
-import 'package:permission_handler/permission_handler.dart';
-import 'package:http/http.dart' as http;
 import 'ui/home.dart';
 import 'ui/chat.dart';
